Added updated server and client script
i made it auto find a tcp port so its not hardcoded so all you have to do is run the server script. also added a recognizer of when the server gets shut down from the client side and disconnects them. fixed problems with padding messages and encryption key and iv transportation